requirements.txt
flask[async]==2.3.3
flask-cors==3.0.10
threema.gateway